src/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.22.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.8.1
aiofiles==24.1.0
motor==3.7.0
pymongo==4.12.0
langchain==0.3.21
langchain-community==0.3.20
unstructured[pdf]==0.17.2
pdf2image==1.17.0
pytesseract==0.3.10
unstructured-inference==0.8.10
transformers==4.41.0
opencv-python-headless==4.7.0.72
tiktoken==0.5.1
tokenizers==0.19
huggingface-hub==0.23.0
accelerate==0.22.0
groq==0.22.0
sentence-transformers==4.1.0
openai==1.75.0
azure-core==1.31.0
azure-ai-inference==1.0.0b9
qdrant-client==1.14.2
voyageai==0.3.2
passlib
python-jose[cryptography]
email-validator
